+++ custom_components/houseplan/frontend/houseplan-card.js
@@ -1,1 +1,1 @@
-globalThis.__HOUSEPLAN_BUILD_FINGERPRINT__="0c61742c98d55aec077478fcee7b8acf52e2ac1b70e2ed23532483b845ca884f";export{d6 as HouseplanCard}from"./houseplan-assets/houseplan-card-BU843YpJ.js";
+globalThis.__HOUSEPLAN_BUILD_FINGERPRINT__="a43476c101e31e0d3e3020d73f02e6dd01f08436f1f404819f58dfbeeebd8f9b";export{d6 as HouseplanCard}from"./houseplan-assets/houseplan-card-BsbMwnTm.js";